Add an ORM skeleton
SQLAlchemy implementations of models from the old Rust/Diesel
implementation, with some updates made possible by SQLAlchemy's richer
support for associations and Postgres semantics.

To create the database, run `python3.6 csgadmin/models.py` on a machine
with a running Postgres instance. The database URL is sourced from a
`.envrc` in the project root.